requirements.txt
Flask==3.0.3
Flask-RESTful==0.3.10
google-api-python-client==2.128.0
google-auth==2.29.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0
seqlog==0.3.31
colorlog==6.8.2